Migrate Gemfile from github-pages to Jekyll 4.2.2
GitHub pages has been stuck for years on an old version of
Jekyll. Updating their convenience tooling to Jekyll 4 would break a
lot of existing sites, so it probably will never happen. See here for
more info: https://github.com/github/pages-gem/issues/651

It turns out that we're already using the Jekyll GitHub Action (I'm
not sure what the reason we were doing that already is), so there's
little reason not to upgrade.

This change also removes a workaround that no longer seems necessary
with the new Jekyll version.