feat: update to Jekyll 4, fix some other things (#56)
* Change _config.yml away from DOS line endings

To sanity check, review on GitHub with `?w=1` url param, or locally
with `git diff --color-words`.

* Migrate Gemfile from github-pages to Jekyll 4.2.2

GitHub pages has been stuck for years on an old version of
Jekyll. Updating their convenience tooling to Jekyll 4 would break a
lot of existing sites, so it probably will never happen. See here for
more info: https://github.com/github/pages-gem/issues/651

It turns out that we're already using the Jekyll GitHub Action (I'm
not sure what the reason we were doing that already is), so there's
little reason not to upgrade.

This change also removes a workaround that no longer seems necessary
with the new Jekyll version.

* Update workflow to cache deps, per action docs

I'm not 100% sure how much this saves, but the docs for the action
mention it in the example:
https://github.com/marketplace/actions/jekyll-actions

* Convert README to asciidoc

Seems reasonable, given this repo is primarily asciidoc-based. :)

* Add instructions for building in a container

* Remove broken include to clear build error

* Enable Liquid pre-processing in adr-process.adoc

See here for more info:
https://github.com/asciidoctor/jekyll-asciidoc/#enabling-liquid-preprocessing

Fixes #49 by causing the template variables to be processed.